new components for DRY, updates to layout and ui.
